refactor(azure): Extract more utils from platform helper
