+++ lib/platform/azure/util.ts
@@ -5,7 +5,9 @@ import {
   PullRequestStatus,
 } from 'azure-devops-node-api/interfaces/GitInterfaces';
 import { logger } from '../../logger';
-import { PrState } from '../../types';
+import { HostRule, PrState } from '../../types';
+import { GitOptions } from '../../types/git';
+import { add } from '../../util/sanitize';
 import { AzurePr } from './types';
 
 export function getNewBranchName(branchName?: string): string {
@@ -127,4 +129,56 @@ export async function streamToString(
     stream.on('error', (err) => reject(err));
   });
   return p;
+}
+
+function toBase64(from: string): string {
+  return Buffer.from(from).toString('base64');
+}
+
+export function getStorageExtraCloneOpts(config: HostRule): GitOptions {
+  let authType: string;
+  let authValue: string;
+  if (!config.token && config.username && config.password) {
+    authType = 'basic';
+    authValue = toBase64(`${config.username}:${config.password}`);
+  } else if (config.token.length === 52) {
+    authType = 'basic';
+    authValue = toBase64(`:${config.token}`);
+  } else {
+    authType = 'bearer';
+    authValue = config.token;
+  }
+  add(authValue);
+  return {
+    '-c': `http.extraheader=AUTHORIZATION: ${authType} ${authValue}`,
+  };
+}
+
+export function max4000Chars(str: string): string {
+  if (str && str.length >= 4000) {
+    return str.substring(0, 3999);
+  }
+  return str;
+}
+
+export function getProjectAndRepo(
+  str: string
+): { project: string; repo: string } {
+  logger.trace(`getProjectAndRepo(${str})`);
+  const strSplit = str.split(`/`);
+  if (strSplit.length === 1) {
+    return {
+      project: str,
+      repo: str,
+    };
+  }
+  if (strSplit.length === 2) {
+    return {
+      project: strSplit[0],
+      repo: strSplit[1],
+    };
+  }
+  const msg = `${str} can be only structured this way : 'repository' or 'projectName/repository'!`;
+  logger.error(msg);
+  throw new Error(msg);
 }
